Updated Series Details Fragment Created Custom Fragments to show schedule and also seasons Improved coded Add Use Case and Repositories

+++ app/src/main/java/com/winterprojects/tvmazeapp/domain/shows/dtos/RatingDto.kt
@@ -3,7 +3,7 @@ package com.winterprojects.tvmazeapp.domain.shows.dtos
 import com.winterprojects.tvmazeapp.domain.shows.models.RatingModel
 
 data class RatingDto(
-    val average: Double?
+    val average: Float?
 )
 
 fun RatingDto.toModel() =
